+++ src/__tests__/__mocks__/api/handlers.js
@@ -1,6 +1,7 @@
 import { rest } from "msw";
 import { items as databases } from "../responses/databases";
 import { items as frameworks } from "../responses/frameworks";
+import { items as programmingLanguages } from "../responses/programmingLanguages";
 
 export const handlers = [
   // Databases
@@ -15,6 +16,13 @@ export const handlers = [
     `${process.env.REACT_APP_BACKEND_URL}/frameworks`,
     (req, res, ctx) => {
       return res(ctx.json(frameworks));
+    }
+  ),
+  // Programming languages
+  rest.get(
+    `${process.env.REACT_APP_BACKEND_URL}/programming-languages`,
+    (req, res, ctx) => {
+      return res(ctx.json(programmingLanguages));
     }
   )
 ];